Removed http2_stream, since it is now unused, and I'd like to repurpose the module name for something else.
Well, technically it's the same thing, but it has to be some kind of
psuedo-fsm that doesn't spawn it's own process, since the only way to
enforce priority is by managing all the streams in one process. also,
multiple processes writing to the same socket will cause all kinds of
trouble.