Removed http2_socket
The logical abstraction is still there, but it made sense to remove the
physical gen_server to reduce the overall number of process messages.

This is a refactor to make it easier to turn http2_streams into real
gen_fsms. I didn't think we'd be able to do that because of stream
prioritization, but now I think that we'd still be able to do stream
priority by slowing the flow of frames to stream fsms, or allowing them
to coordinate with the connection fsm. Either way, it's still optional,
so a ways off, if ever.