Introduces h2_stream_set
h2_stream_set is a data structure built to track all streams in an
http/2 connection. That's right, all 2^31-1 of them. It works with a
smaller set in memory, and treats the set of connections initiated by
each peer separately. Since we need to track the number of active
streams in each direction.