Fixed a fatal flaw in flow control math
The first problem was a case clause that was calling rst_stream/3, but
not being the return of the function. So the connection state never got
the memo that the stream was reset.

Also there was a problem with automatic flow control math. The problem
is simple to explain.

Say your stream's initial flow control window is 65K. It's a fair
assumption, since that's the default. Say we send that stream a frame
that's 16K. Also a default size. Since we're calculating flow control
windows, the new window size would be 65K - 16K = 49K. Great. But with
automatic flow control, we're immediately sending a WINDOW_UPDATE of
16K. This tells the sender that we're willing to receive 65K again, but
on the receiving side, we weren't incrementing the window, so we stayed
at 49K. Every frame would just lower what we thought we could recieve,
but cause no change on the client side.