Fix check of PING ACK flag
Sending a PING from a chatterbox client to a chatterbox server results
in an infinite PING loop. This is because the check for the PING ACK
flag is incorrect (it is missing the 'H' variable, leading to the ACK
flag being seen always as off).

This commit fixes the bug.