Add h2_connection:update_settings/2
In a server context, it is often necessary to update settings
dynamically. A realistic example of this is when changing MCS, depending
on connection parameters. This is something that APNS does, the behavior
of which is mimicked in the APNS simulator project[1].

This commit adds a call that enables the client to update settings.

[1]: https://github.com/SilentCircle/apns_erl_sim