wait to encode trailers and updating hpack dynamic table until sending
Trailer encoding and sending is now done in h2_connection process
when the trailers are ready to be sent (body has been sent). This
fixes a bug where headers or trailers of a later stream could go out
before a set of trailers that were already encoded but not sent and
so the hpack dynamic table would be out of sync between client/server.