Don't set END_STREAM if there are trailers to send
Calling h2_connection:send_trailers ends up setting body_complete
fo the stream to true. But it can be that the h2_connection is
still in the process of sending data if a larger data has been
given to send_body. Then the END_STREAM flag is falsely set already
at the last DATA frame.

This commit fixes the issue by not setting the END_STREAM flag on
the DATA frame if there are trailers defined to be sent.