Improve performance by not waiting in sock:recv
A 1ms timeout was used in sock:recv. This has a big impact on
performance. By immediately continuing processing if no data is
available, the number of messages that can be handled per second
improves from less than 200 per connection to thousands or tens of
thousands (depending on resources).

This fix reveals also a small issue with handling frames in a stream
that is about to close. Test half_closed_remote_sends_headers of the
http2_spec_5_1_SUITE starts occasionally failing. This is not a new
issue introduced by this commit but something that now occurs more
frequently due to the faster handling. Previously in this test, the
server side had actually always already reached closed state before
handling the extra frame of the test, not in half_closed_remote as
the test says. With this commit, the state might still be
half_closed_remote. This reveals a design issue that when a h2_stream
active but already has the messages in the gen_statem queue that will
end up closing it, then the h2_connection will still add new actions
to the queue since from its view the stream is active. Then when the
h2_stream reaches the message that closes it, it will close the
stream and discard the rest of the queue. In this test case the
falsely sent HEADERS frame goes unanswered, even though a RST_STREAM
reply should have been sent.

Due to the above, the half_closed_remote_sends_headers test has been
replaced with a closed_receives_headers which tests the same as the
previous test was actually testing due to slower recv handling. The
half_closed_remote_sends_headers has been left as disabled.