Send WINDOW_UPDATE on SETTINGS acknowledgement
When using a window size greater than the default 65535, the window
size of the connection needs to be updated by sending a WINDOW_UPDATE
message. According to RFC7540, only the stream flow-control window
is updated directly by the SETTINGS message, the connection
flow-control window needs to be updated with a WINDOW_UPDATE message.