Add support for setting TCP_USER_TIMEOUT
TCP_USER_TIMEOUT enables detecting lost connectivity and
shutting down connections in a timely manner. It provides the user
with control on how quickly to give up on connections without
connectivity, for example in order to failover to an alternative
connection.

The change uses socket options of type raw, set for a Linux system.
It will probably only work on Linux and should not be used on other
platforms.