prevent crash on window update with closed socket
The window update window update sending was assumed to always
succeed. This is not the case when the socket has been closed.
For example, when there are messages in the gen_statem message
queue then operations can be run on an already closed socket.

The fix ignores the socket error and just continues, similar to
what other socket send operations seem to do in this module.
Eventually the tcp_closed/tcp_error event should be handled by
handle_event and the connection go down.